requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
gunicorn>=20.0.4,<20.1
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
django-request-logging==0.7.5
djangorestframework-simplejwt==5.2.0
idna==3.3
PyJWT==2.4.0
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.10